move from resource to the rest publisher area
